pyproject.toml
[project]
name = "template"
version = "1.0.0"
description = "Bot template"
authors = [
    {name = "Yuriy Chebyshev", email = "bomzheg@gmail.com"},
]
license = {file = "LICENSE"}
readme = "README.md"
requires-python = ">=3.13,<3.15"
dependencies = [
    "pyyaml>=6.0,<7.0",
    "python-dateutil>=2.8.2,<3.0",
    "colorlog>=6.7.0,<7.0",
    "aiogram>=3.20.0,<3.21.0",
    "pydantic>=2.11,<3.0",
    "aiogram_dialog[tools]>=2.1,<2.2",
    "sqlalchemy[asyncio]>=2.0,<3.0",
    "alembic>=1.10.0,<2.0",
    "asyncpg>=0.27.0,<1.0",
    "ujson>=5.5.0,<6.0",
    "redis>=5.0.1,<6.0",
    "adaptix==3.0.0b9",
    "fastapi==0.111.1",
    "uvicorn>=0.30.0,<0.31.0",
    "greenlet<4.0.0",
    "dishka==1.5.2",
    "asgi-monitor>=0.4.0,<0.5.0",
]

[project.optional-dependencies]
dev = [
    "locust>=2.15.1,<3",
]
test = [
    "pytest>=7.2.0,<8.0",
    "pytest-ordering>=0.6,<0.7",
    "pytest-asyncio>=0.18.3,<0.19",
    "pytest-cov",
    "testcontainers>=4.5.1,<5.0",
    "httpx>=0.23.0,<0.24.0",
    "coverage[toml]",
    "ruff>=0.2.1,<0.3",
    "types-python-dateutil>=2.8.19,<3.0",
    "types-redis>=4.3.21,<5.0",
    "types-PyYAML>=6.0.12,<7.0",
    "mypy>=1.1.1,<2.0",
    "asgi-lifespan>=2.1.0,<3.0",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]

[tool.coverage.run]
branch = true
relative_files = true
include = ["app/**"]
omit = [
    ".*",
    "*/site-packages/*",
    "*/__main__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "class .*\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
    "(typing\\.)?assert_never",
]

[tool.black]
line-length = 99
target-version = ['py313', 'py314']
include_trailing_comma = true
exclude = '''
(
    \.eggs
  | \.git
  | \.tox
  | build
  | dist
  | venv
  | docs
  | .idea
)
'''

[tool.ruff]
line-length = 99

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D",  # docstrings
    "ANN101", "ANN102",  # annotations for self and cls
    "EM101",  # forbidden exceptions with string literals
    "TRY003",  # forbidden long text messages for create exceptions
    "FA100", "FA102",  # from future import annotation related
    "TD002", "TD003", "TD004", "TD005",  # useless TO DO rules
    "COM812", "ISC001",  # conflicted rules with ruff format
    "TCH001", "TCH002", "TCH003",  # required TYPE_CHECKING block
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = [
    "S101",
]
"**/migrations/versions/*" = ["N999"]

[tool.mypy]
no_implicit_optional = true
warn_unused_ignores = true
show_error_codes = true
exclude = ['venv', '.venv']

[[tool.mypy.overrides]]
module = "tests.*"
disable_error_code = ["func-returns-value"]


[[tool.mypy.overrides]]
module = [
    "apscheduler.*",
    "testcontainers.*",
    "diagrams.*",
]
ignore_missing_imports = true
